Fixing PTRArchive (with updated URL)
Note: This commit resolves an issue with the PTRArchive collector.
There is another pull request that exists for this issue, but I believe
it’s hitting a URL that’s since been changed.

However, on a larger scale, the owner of PTRArchive seems to not be a
huge fan of tools querying the data, and seems to be indiscriminately
labeling all applications (even those with legitimate utility for
penetration testers) 'black hat tools'. In the case of Sublist3r, this
source was removed as a source after tampering with returned data or
blocking wholesale all requests. I believe this may actually be why the
URL has shifted between my pull request and the previous. Perhaps this
collector should be removed?